backend/submit-application/requirements.txt
psycopg2-binary==2.9.9
pydantic==2.5.0
email-validator==2.1.0